refactor: step3 피드백 반영
- Session 수강신청 Enrollment로 위임
