+++ src/main/java/nextstep/courses/domain/Session.java
@@ -1,6 +1,5 @@
 package nextstep.courses.domain;
 
-import nextstep.courses.exception.SessionException;
 import nextstep.payments.domain.Payment;
 import nextstep.users.domain.NsUser;
 
@@ -17,15 +16,7 @@ public class Session {
 
     private final SessionPeriod sessionPeriod;
 
-    private final SessionPrice sessionPrice;
-
-    private final SessionState sessionState;
-
-    private final SessionType sessionType;
-
-    private final SessionUsers sessionUsers;
-
-    private final SessionUserCount sessionUserCount;
+    private final Enrollment enrollment;
 
     public Session(Long id) {
         this(id, 1L,
@@ -34,65 +25,25 @@ public class Session {
                         LocalDate.of(2023, 1, 1),
                         LocalDate.of(2023, 1, 30)
                 ),
-                new SessionPrice(1000),
-                SessionState.PREPARE,
-                SessionType.PAID,
-                new SessionUserCount(0, 0)
+                new Enrollment(
+                        new SessionPrice(1000),
+                        SessionState.PREPARE,
+                        SessionType.PAID,
+                        new SessionUserCount(0, 0)
+                )
         );
     }
 
-    public Session(Long id, Long courseId, SessionImage sessionImage, SessionPeriod sessionPeriod, SessionPrice sessionPrice, SessionState sessionState, SessionType sessionType, SessionUserCount sessionUserCount) {
-        this(id, courseId, sessionImage, sessionPeriod, sessionPrice, sessionState, sessionType, new SessionUsers(), sessionUserCount);
-    }
-
-    public Session(Long id, Long courseId, SessionImage sessionImage, SessionPeriod sessionPeriod, SessionPrice sessionPrice, SessionState sessionState, SessionType sessionType, SessionUsers sessionUsers, SessionUserCount sessionUserCount) {
+    public Session(Long id, Long courseId, SessionImage sessionImage, SessionPeriod sessionPeriod, Enrollment enrollment) {
         this.id = id;
         this.courseId = courseId;
         this.sessionImage = sessionImage;
         this.sessionPeriod = sessionPeriod;
-        this.sessionPrice = sessionPrice;
-        this.sessionState = sessionState;
-        this.sessionType = sessionType;
-        this.sessionUsers = sessionUsers;
-        this.sessionUserCount = sessionUserCount;
+        this.enrollment = enrollment;
     }
 
     public void register(NsUser user, Payment payment) {
-        validateState();
-        validateType();
-        validatePriceEqualPayment(payment);
-        sessionUsers.addUser(user);
-        sessionUserCount.plusUserCount();
-    }
-
-    private void validatePriceEqualPayment(Payment payment) {
-        if (isPaid()) {
-            sessionPrice.isSameBy(payment);
-        }
-    }
-
-    private void validateState() {
-        if (!isOpen()) {
-            throw new SessionException("현재 강의가 모집중이지 않아 수강 신청을 할 수가 없습니다.");
-        }
-    }
-
-    private boolean isOpen() {
-        return sessionState.isOpen();
-    }
-
-    private void validateType() {
-        if (isPaid()) {
-            sessionUserCount.validateMaxUserCount();
-        }
-    }
-
-    private boolean isPaid() {
-        return sessionType == SessionType.PAID;
-    }
-
-    public int userCount() {
-        return sessionUserCount.userCount();
+        enrollment.enroll(user, payment);
     }
 
     @Override
